chore: 초대 중인 멤버 제외, 리팩토링
