+++ functions/db/memberDB.js
@@ -215,6 +215,24 @@ const getAllInvitedTeamIdList = async (client, userId) => {
   return convertSnakeToCamel.keysToCamel(rows);
 };
 
+const getMemberByTeamId = async (client, teamId) => {
+  const { rows } = await client.query(
+    `
+      SELECT u.id, u.profile_id, u.image, u.name, m.is_host
+      FROM "member" m JOIN "user" u
+      ON m.user_id = u.id
+      WHERE m.team_id = $1
+      AND m.is_deleted = false
+      AND u.is_deleted = false
+      AND m.is_confirmed = true
+      ORDER BY m.is_host DESC
+      `,
+    [teamId],
+  );
+
+  return convertSnakeToCamel.keysToCamel(rows);
+};
+
 module.exports = {
   getAllTeamByUserId,
   getAllTeamMemberByTeamId,
@@ -229,4 +247,5 @@ module.exports = {
   updateNewHost,
   getInvitedTeamIdList,
   getAllInvitedTeamIdList,
+  getMemberByTeamId,
 };
